+++ DataObjects/JobConfiguration/JobTarget.cs
@@ -8,7 +8,7 @@ namespace AppDynamics.Dexter
         public string UserName { get; set; }
         public string UserPassword { get; set; }
         public string Application { get; set; }
-        public int ApplicationID { get; set; }
+        public long ApplicationID { get; set; }
         public bool NameRegex { get; set; }
         public JobTargetStatus Status { get; set; }
